fix: adiciona tela do filme
